+++ app/gui/trading/__init__.py
@@ -2,6 +2,7 @@ from .best_buyer_tab import BestBuyerTab
 from .commodities_tab import CommoditiesTab
 from .en_route_tab import EnRouteTab
 from .shops_tab import ShopsTab
+from .trade_routes_tab import TradeRoutesTab
 from .trading_tab import TradingTab
 from .uex_trading_tab import UEXTradingTab
 
@@ -11,6 +12,7 @@ __all__ = [
     "CommoditiesTab",
     "EnRouteTab",
     "ShopsTab",
+    "TradeRoutesTab",
     "TradingTab",
     "UEXTradingTab",
 ]